pyproject.toml
[project]
name = "landscaping-erm"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "dbt-bigquery>=1.11.1",
    "google-cloud-bigquery>=3.41.0",
    "pyodbc>=5.0.0",
    "pyarrow>=15.0.0",
]
